+++ CHANGELOG.md
@@ -94,6 +94,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added pdr-cleanup task into cumulus core from ASDC
 - **CUMULUS-4563**
   - Added a Github action to generate requirements.txt files from coreified uv.lock files
+- **CUMULUS-4428**
+  - Added add-input-granules task
+- **CUMULUS-4426**
+  - Added unit tests for add-input-granules task
 
 ### Changed